importYelpRestaurants function pulls down data from Yelp
